requirements.txt
chainer==1.17.0
cycler==0.10.0
dask==0.12.0
decorator==4.0.10
filelock==2.0.7
lmdb==0.92
matplotlib==1.5.3
networkx==1.11
nose==1.3.7
numpy==1.11.2
Pillow==3.4.2
protobuf==3.1.0.post1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
scikit-image==0.12.3
scipy==0.18.1
six==1.10.0
toolz==0.8.0
